+++ spend_ya_telegram/__init__.py
@@ -1,0 +1,1 @@
+default_app_config = 'spend_ya_telegram.apps.SpendYaTelegramConfig'
