+++ spend_ya_telegram/urls.py
@@ -1,0 +1,8 @@
+from django.conf.urls import url
+
+import views
+
+
+urlpatterns = [
+    url(r'^(?P<telegram_token>\w+)', views.telehook, name='hook')
+]
